feat: support basic loops, br_if

+++ crates/tinywasm/src/runtime/value.rs
@@ -10,7 +10,7 @@ pub struct RawWasmValue(u64);
 
 impl Debug for RawWasmValue {
     fn fmt(&self, f: &mut core::fmt::Formatter) -> core::fmt::Result {
-        write!(f, "RawWasmValue({})", self.0)
+        write!(f, "RawWasmValue({})", self.0 as i64) // cast to i64 so at least negative numbers for i32 and i64 are printed correctly
     }
 }
